requirements.txt
python-telegram-bot==21.6
anthropic>=0.40.0
httpx>=0.27.0
gTTS>=2.5.0
